Refactor CRM API structure: Implement SRP by breaking down the API into individual files for each function, enhancing clarity and maintainability. Update README to reflect new API structure and usage. Adjust import paths in hooks for consistency. Correct minor text formatting in LeadDetail component.
